fix up breathe steps bug

+++ src/effects/spectrum_cycle.cc
@@ -7,7 +7,7 @@
 #include "spectrum_cycle.hh"
 #include "color_cycling.hh"
 
-SpectrumCycle::SpectrumCycle(Region &region, Color baseColor, double delay) :
+SpectrumCycle::SpectrumCycle(Region &region, Color baseColor, int16_t delay) :
         Effect(region), base(baseColor), active(baseColor), delay(delay) {
 
 }
